+++ src/main/java/ru/shipilov/diplom/logic/utils/Stat.java
@@ -1,7 +1,7 @@
 package ru.shipilov.diplom.logic.utils;
 
 public class Stat{
-    public static int Sturges(int count){
+    public static int Sturges(Long count){
         return (int)(1+3.322*Math.log(count));
     }
 }
